% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSBLUP.R
\name{cv.cal_ab}
\alias{cv.cal_ab}
\title{Cross validation calculation for accuracy}
\usage{
cv.cal_ab(
  phe_ab,
  trait,
  M1,
  M2,
  M3,
  cov,
  a,
  b,
  crv.num,
  crv.rep.num,
  stas.phe.col = NULL,
  stas.type,
  stas.fn,
  ...
)
}
\arguments{
\item{phe_ab}{phenotype data.frame}

\item{trait}{analysis trait}

\item{M1}{pedigree kinship matrix}

\item{M2}{genomic matrix}

\item{M3}{omics matrix}

\item{a}{parameter \verb{\\alpha}}

\item{b}{parameter \verb{\\beta}}

\item{crv.num}{fold number of validation}

\item{crv.rep.num}{repeat times in each fold, also as random seed}

\item{stas.phe.col}{target trait when calculate the indicator, default as @inheritParams trait}

\item{stas.type}{indicator when optimize the fusion matrix}

\item{stas.fn}{custom calculated indicators}

\item{...}{passed parameters for other function}
}
\value{
list
}
\description{
Cross validation calculation for accuracy
}
