% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSBLUP.R
\name{FSBLUP.Bisection}
\alias{FSBLUP.Bisection}
\title{bisection optimize points from grid_search returns}
\usage{
FSBLUP.Bisection(
  acc_df,
  fn,
  max_iter = 10,
  threshold = 1e-04,
  phe_ab,
  trait,
  M1,
  M2,
  M3,
  cov,
  crv.num,
  crv.rep.num,
  stas.phe.col,
  stas.type,
  stas.fn,
  train.id = NULL,
  test.id = NULL,
  ...
)
}
\arguments{
\item{acc_df}{grid_search returns}

\item{fn}{function to calculate cor}

\item{max_iter}{max bisection iterations}

\item{threshold}{stop if reached between two iterations}

\item{phe_ab}{phenotype data.frame}

\item{trait}{analysis trait}

\item{M1}{pedigree kinship matrix}

\item{M2}{genomic matrix}

\item{M3}{omics matrix}

\item{crv.num}{fold number of validation}

\item{crv.rep.num}{repeat times in each fold, also as random seed}

\item{stas.phe.col}{target trait when calculate the indicator, default as @inheritParams trait}

\item{stas.type}{indicator when optimize the fusion matrix}

\item{stas.fn}{custom calculated indicators}

\item{train.id}{reference population id}

\item{test.id}{validation population id}

\item{...}{passed parameters for other function}
}
\value{
the best points
}
\description{
bisection optimize points from grid_search returns
}
