% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSBLUP.R
\name{FSBLUP.CombineMat}
\alias{FSBLUP.CombineMat}
\title{Combine matrix}
\usage{
FSBLUP.CombineMat(phe_c, M1, M2, M3, a, b)
}
\arguments{
\item{phe_c}{phenotype data.frame}

\item{M1}{A pedigree matrix}

\item{M2}{G genomic matrix}

\item{M3}{O omics matrix}

\item{a}{weight for M3}

\item{b}{weight for M2}
}
\value{
matrix
}
\description{
Combine matrix
}
