% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSBLUP.R
\name{FSBLUP.GridSearch}
\alias{FSBLUP.GridSearch}
\title{grid search for a given range of points}
\usage{
FSBLUP.GridSearch(
  fn,
  x = c(0, 1),
  y = c(0, 1),
  point_num = 25,
  phe_ab,
  trait,
  M1,
  M2,
  M3,
  cov,
  crv.num,
  crv.rep.num,
  stas.phe.col,
  stas.type,
  stas.fn,
  train.id = NULL,
  test.id = NULL,
  ...
)
}
\arguments{
\item{fn}{function to calculate indicator}

\item{x}{search ranges, e.g. (0,1)}

\item{y}{search ranges, e.g. (0,1)}

\item{point_num}{search density on two directions, steps is \code{sqrt(point_num)}}

\item{phe_ab}{phenotype data.frame}

\item{trait}{analysis trait}

\item{M1}{pedigree kinship matrix}

\item{M2}{genomic matrix}

\item{M3}{omics matrix}

\item{crv.num}{fold number of validation}

\item{crv.rep.num}{repeat times in each fold, also as random seed}

\item{stas.phe.col}{target trait when calculate the indicator, default as @inheritParams trait}

\item{stas.type}{indicator when optimize the fusion matrix}

\item{stas.fn}{custom calculated indicators}

\item{train.id}{reference population id}

\item{test.id}{validation population id}

\item{...}{passed parameters for other function}

\item{a}{parameter \verb{\\alpha}}

\item{b}{parameter \verb{\\beta}}
}
\value{
a data.frame induces search points and cor
}
\description{
grid search for a given range of points
}
