% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSBLUP.R
\name{FSBLUP.Mix}
\alias{FSBLUP.Mix}
\title{solve the mix model and calculate ebvs}
\usage{
FSBLUP.Mix(phe, trait_col, k)
}
\arguments{
\item{phe}{phenotype data.frame}

\item{trait_col}{trait col num in phe}

\item{k}{kinship matrix}
}
\description{
solve the mix model and calculate ebvs
}
