% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSBLUP.R
\name{FSBLUP.Omatrix}
\alias{FSBLUP.Omatrix}
\title{Calculate omics similarity matrix}
\usage{
FSBLUP.Omatrix(data = NULL, missing.value = NA, FPKM.qc = FALSE)
}
\arguments{
\item{data}{omics data}

\item{missing.value}{missing value in data, will substitute to 0}

\item{FPKM.qc}{if omics data is transcriptomic fprk format data, will filter transcripts which fpkm < 0.1 in more than 95\% individuals}
}
\value{
matrix
}
\description{
Calculate omics similarity matrix
}
