% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSBLUP.R
\name{FSBLUP}
\alias{FSBLUP}
\title{Fusion similarity matrix calculation}
\usage{
FSBLUP(
  phe = NULL,
  M1 = NULL,
  M2 = NULL,
  M3 = NULL,
  pedi = NULL,
  snp = NULL,
  omic = NULL,
  FPKM.qc = TRUE,
  cov = NULL,
  trait_col = NULL,
  po.crv.num = 5,
  po.crv.rep.num = 2,
  po.year = NULL,
  po.ngen = NULL,
  po.gs.point_num = 25,
  po.bi.max_iter = 10,
  po.bi.threshold = 0.0001,
  stas.type = "cor",
  stas.fn = NULL,
  stas.phe.col = NULL,
  return.matrix = FALSE
)
}
\arguments{
\item{phe}{pheno data.frame}

\item{M1}{pedigree kinship matrix}

\item{M2}{genomic matrix}

\item{M3}{omics matrix}

\item{pedi}{pedigree data.frame, id|sire|dam}

\item{snp}{snp data.frame, individuals(row) * snps(col)}

\item{omic}{omic data.frame, individuals(row) * omic features(col)}

\item{trait_col}{analysis trait in the \code{phe}, col number or name}

\item{po.crv.num}{fold number of cross validation in parameter optimization procedure}

\item{po.crv.rep.num}{repeat times of cross validation in parameter optimization procedure}

\item{po.year}{year variable col of time order validation in parameter optimization procedure}

\item{po.ngen}{ngen variable col of time order validation in parameter optimization procedure}

\item{po.gs.point_num}{grid search point number, default 25}

\item{po.bi.max_iter}{bisection maximum iterations}

\item{po.bi.threshold}{bisection stop  threshold between two iterations}

\item{stas.type}{indicator when optimize the fusion matrix}

\item{stas.fn}{custom calculated indicators}

\item{stas.phe.col}{target trait when calculate the indicator, default as @inheritParams trait_col}

\item{return.matrix}{default \code{T} return fusion matrix, \code{F} return predicted EBVs}
}
\value{
matrix
}
\description{
Fusion similarity matrix calculation
}
