% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSBLUP.R
\name{FSBLUP.TimeV}
\alias{FSBLUP.TimeV}
\title{Perform time order validation in reference population to find best parameters}
\usage{
FSBLUP.TimeV(
  phe_cv,
  M1,
  M2,
  M3,
  trait_col,
  cov,
  train.id,
  test.id,
  gs.point_num,
  bi.max_iter,
  bi.threshold,
  stas.phe.col,
  stas.type,
  stas.fn
)
}
\arguments{
\item{phe_cv}{pheno data.frame}

\item{M1}{pedigree kinship matrix}

\item{M2}{genomic matrix}

\item{M3}{omics matrix}

\item{trait_col}{analysis trait col number or name}

\item{train.id}{reference population id}

\item{test.id}{validation population id}

\item{gs.point_num}{grid search point number, default 25}

\item{bi.max_iter}{bisection maximum iterations}

\item{bi.threshold}{bisection stop  threshold between two iterations}

\item{stas.phe.col}{target trait when calculate the indicator, default as @inheritParams trait_col}

\item{stas.type}{indicator when optimize the fusion matrix}

\item{stas.fn}{custom calculated indicators}
}
\value{
list
}
\description{
Perform time order validation in reference population to find best parameters
}
