% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSBLUP.R
\name{tv.cal_ab}
\alias{tv.cal_ab}
\title{Time order calculation for accuracy}
\usage{
tv.cal_ab(
  phe_ab,
  trait,
  M1,
  M2,
  M3,
  cov,
  a,
  b,
  RP.id = NULL,
  VP.id = NULL,
  stas.phe.col = NULL,
  stas.type,
  stas.fn,
  ...
)
}
\arguments{
\item{phe_ab}{phenotype data.frame}

\item{trait}{analysis trait}

\item{M1}{pedigree kinship matrix}

\item{M2}{genomic matrix}

\item{M3}{omics matrix}

\item{a}{parameter \verb{\\alpha}}

\item{b}{parameter \verb{\\beta}}

\item{RP.id}{reference population id}

\item{VP.id}{validation population id}

\item{stas.phe.col}{target trait when calculate the indicator, default as @inheritParams trait}

\item{stas.type}{indicator when optimize the fusion matrix}

\item{stas.fn}{custom calculated indicators}

\item{...}{passed parameters for other function}
}
\value{
list
}
\description{
Time order calculation for accuracy
}
